refactor: split internal/user/manager.go by domain
Pure move: manager.go (825 lines, 34 declarations) split into
concern-named sibling files (manager_users, manager_calls,
manager_persist, manager_purge, manager_auth, manager_stats),
following the same-package sibling-file convention used by
internal/config/config_*.go and internal/message/manager_*.go. Only
import blocks changed; no behavior, exported API, or test changes.
