+++ internal/user/manager.go
@@ -12,7 +12,6 @@ import (
 	"time"
 
 	"golang.org/x/crypto/bcrypt" // Import bcrypt
-	"golang.org/x/crypto/ssh"
 )
 
 // Predefined errors for user management
@@ -193,623 +192,6 @@ func (um *UserMgr) determineNextUserID() { // Receiver uses renamed type
 	um.nextUserID = maxID + 1
 	slog.Debug("determined next user ID", "id", um.nextUserID)
 	um.mu.Unlock()
-}
-
-// loadCallHistory loads the call history events from JSON.
-// Now uses um.dataPath internally.
-func (um *UserMgr) loadCallHistory() error {
-	filePath := filepath.Join(um.dataPath, callHistoryFile) // Use stored dataPath
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("call history file not found; starting empty", "file", callHistoryFile)
-			return nil // Not an error if the file doesn't exist yet
-		}
-		return fmt.Errorf("failed to read %s: %w", callHistoryFile, err)
-	}
-
-	data = StripUTF8BOM(data)
-	if len(data) == 0 {
-		return nil // Empty file is okay
-	}
-
-	um.mu.Lock() // Lock before modifying internal slice
-	defer um.mu.Unlock()
-	// Ensure slice exists
-	if um.callHistory == nil {
-		um.callHistory = make([]CallRecord, 0, callHistoryLimit)
-	}
-	if err := json.Unmarshal(data, &um.callHistory); err != nil {
-		return fmt.Errorf("failed to unmarshal %s: %w", callHistoryFile, err)
-	}
-	// Ensure capacity and length limits are respected after loading
-	if len(um.callHistory) > callHistoryLimit {
-		startIdx := len(um.callHistory) - callHistoryLimit
-		um.callHistory = um.callHistory[startIdx:]
-	}
-	slog.Debug("loaded call history records", "count", len(um.callHistory), "file", callHistoryFile)
-	return nil
-}
-
-// loadNextCallNumber loads the next call number from its dedicated JSON file.
-func (um *UserMgr) loadNextCallNumber() error {
-	filePath := filepath.Join(um.dataPath, callNumberFile)
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			slog.Info("call number file not found; starting from 1", "file", callNumberFile)
-			// Keep the default um.nextCallNumber = 1
-			return nil // Not an error if the file doesn't exist
-		}
-		return fmt.Errorf("failed to read %s: %w", callNumberFile, err)
-	}
-
-	data = StripUTF8BOM(data)
-	if len(data) == 0 {
-		slog.Warn("call number file is empty; starting from 1", "file", callNumberFile)
-		return nil // Empty file, use default
-	}
-
-	um.mu.Lock() // Lock before modifying
-	defer um.mu.Unlock()
-	if err := json.Unmarshal(data, &um.nextCallNumber); err != nil {
-		// If unmarshal fails, log and keep the default
-		slog.Warn("failed to unmarshal call number file; starting from 1", "file", callNumberFile, "error", err)
-		um.nextCallNumber = 1
-		return nil // Don't return error, just use default
-	}
-
-	slog.Debug("loaded next call number", "number", um.nextCallNumber, "file", callNumberFile)
-	return nil
-}
-
-// saveCallHistoryLocked saves the current callHistory slice to JSON (assumes lock is held).
-// Now uses um.dataPath internally.
-func (um *UserMgr) saveCallHistoryLocked() error {
-	if um.callHistory == nil {
-		// Avoid marshaling nil slice, treat as empty
-		um.callHistory = make([]CallRecord, 0)
-	}
-	data, err := json.MarshalIndent(um.callHistory, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal call history: %w", err)
-	}
-
-	filePath := filepath.Join(um.dataPath, callHistoryFile) // Use stored dataPath
-	if err := os.WriteFile(filePath, data, 0644); err != nil {
-		return fmt.Errorf("failed to write %s: %w", callHistoryFile, err)
-	}
-
-	// Also save the next call number (atomically with history? separate file is simpler for now)
-	if err := um.saveNextCallNumberLocked(); err != nil {
-		// Log error but don't fail the history save if number save fails
-		slog.Error("failed to save next call number", "error", err)
-	}
-
-	return nil
-}
-
-// saveNextCallNumberLocked saves the current nextCallNumber to its JSON file (assumes lock is held).
-func (um *UserMgr) saveNextCallNumberLocked() error {
-	data, err := json.Marshal(um.nextCallNumber) // Simple marshal, no indent needed
-	if err != nil {
-		return fmt.Errorf("failed to marshal next call number: %w", err)
-	}
-
-	filePath := filepath.Join(um.dataPath, callNumberFile)
-	if err := os.WriteFile(filePath, data, 0644); err != nil {
-		return fmt.Errorf("failed to write %s: %w", callNumberFile, err)
-	}
-	return nil
-}
-
-// saveUsersLocked performs the actual saving without acquiring locks.
-// Uses um.path (which should point to data/users.json)
-func (um *UserMgr) saveUsersLocked() error { // Receiver uses renamed type
-	// Convert map back to slice for saving as JSON array.
-	// Clear LegacyUsername before marshaling so the old "username" key is not written back.
-	usersList := make([]*User, 0, len(um.users))
-	for _, user := range um.users {
-		if user.LegacyUsername != "" {
-			cp := *user
-			cp.LegacyUsername = ""
-			usersList = append(usersList, &cp)
-		} else {
-			usersList = append(usersList, user)
-		}
-	}
-
-	data, err := json.MarshalIndent(usersList, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal users slice: %w", err)
-	}
-
-	// Ensure the directory exists before writing the file
-	dir := filepath.Dir(um.path)
-	if err := os.MkdirAll(dir, 0750); err != nil { // Use 0750 for permissions
-		return fmt.Errorf("failed to create directory %s: %w", dir, err)
-	}
-
-	// WriteFile ensures atomic write (usually via temp file)
-	if err = os.WriteFile(um.path, data, 0600); err != nil {
-		return fmt.Errorf("failed to write users file %s: %w", um.path, err) // Include path in error
-	}
-	return nil
-}
-
-// SaveUsers saves the current user data to the JSON file (acquires lock).
-func (um *UserMgr) SaveUsers() error { // Receiver uses renamed type
-	um.mu.Lock()
-	defer um.mu.Unlock()
-	return um.saveUsersLocked()
-}
-
-// UpdateUserByID updates a user looked up by their stable ID, safely re-keying
-// the internal map when the handle has changed. Use this in any flow that may
-// rename a user's handle (e.g., admin editor); for all other updates prefer
-// UpdateUser.
-func (um *UserMgr) UpdateUserByID(u *User) error {
-	if u == nil {
-		return fmt.Errorf("cannot update nil user")
-	}
-	u.Handle = strings.TrimSpace(u.Handle)
-	if u.Handle == "" {
-		return fmt.Errorf("handle cannot be blank")
-	}
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Locate existing map entry by stable ID.
-	var oldKey string
-	for k, existing := range um.users {
-		if existing.ID == u.ID {
-			oldKey = k
-			break
-		}
-	}
-	if oldKey == "" {
-		return ErrUserNotFound
-	}
-
-	newKey := strings.ToLower(u.Handle)
-	var originalEntry *User
-	rekeyed := false
-	if newKey != oldKey {
-		// Handle changed — ensure no collision with a different user.
-		if existing, exists := um.users[newKey]; exists && existing.ID != u.ID {
-			return ErrHandleExists
-		}
-		originalEntry = um.users[oldKey] // save for rollback
-		delete(um.users, oldKey)
-		rekeyed = true
-	}
-
-	userCopy := *u
-	um.users[newKey] = &userCopy
-	if err := um.saveUsersLocked(); err != nil {
-		// Rollback in-memory map to match what is still on disk.
-		delete(um.users, newKey)
-		if rekeyed {
-			um.users[oldKey] = originalEntry
-		}
-		return err
-	}
-	return nil
-}
-
-// UpdateUser copies the modified user back into the internal map and saves to disk.
-// Use this instead of SaveUsers when you have modified a user copy obtained from
-// GetUser or Authenticate and need those changes persisted.
-func (um *UserMgr) UpdateUser(u *User) error {
-	if u == nil {
-		return fmt.Errorf("cannot update nil user")
-	}
-	um.mu.Lock()
-	defer um.mu.Unlock()
-	lowerHandle := strings.ToLower(u.Handle)
-	if _, exists := um.users[lowerHandle]; !exists {
-		return ErrUserNotFound
-	}
-	// Create a defensive copy to prevent external mutations from bypassing locks
-	userCopy := *u
-	previous := um.users[lowerHandle]
-	um.users[lowerHandle] = &userCopy
-	if err := um.saveUsersLocked(); err != nil {
-		// Restore the previous entry so the cache never serves a value that
-		// failed to reach disk, matching AddUser and UpdateUserByID.
-		um.users[lowerHandle] = previous
-		return err
-	}
-	return nil
-}
-
-// LogAdminActivity logs an administrative action to the activity log file
-func (um *UserMgr) LogAdminActivity(logEntry AdminActivityLog) error {
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Load existing logs
-	logPath := filepath.Join(filepath.Dir(um.path), adminLogFile)
-	var logs []AdminActivityLog
-
-	// Try to load existing logs
-	if data, err := os.ReadFile(logPath); err == nil {
-		data = StripUTF8BOM(data)
-		_ = json.Unmarshal(data, &logs) // Ignore errors, start fresh if corrupt
-	}
-
-	// Add new entry
-	logEntry.ID = len(logs) + 1
-	logEntry.Timestamp = time.Now()
-	logs = append(logs, logEntry)
-
-	// Keep only recent entries (prevent file from growing indefinitely)
-	if len(logs) > adminLogLimit {
-		logs = logs[len(logs)-adminLogLimit:]
-	}
-
-	// Save logs
-	data, err := json.MarshalIndent(logs, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal admin logs: %w", err)
-	}
-
-	if err := os.WriteFile(logPath, data, 0600); err != nil {
-		return fmt.Errorf("failed to write admin log: %w", err)
-	}
-
-	return nil
-}
-
-// Authenticate checks handle and compares password hash.
-// Handle lookup is case-insensitive.
-// Returns: (user, success)
-func (um *UserMgr) Authenticate(handle, password string) (*User, bool) {
-	lowerHandle := strings.ToLower(handle)
-
-	um.mu.RLock()
-	user, exists := um.users[lowerHandle]
-	if !exists {
-		um.mu.RUnlock()
-		return nil, false
-	}
-	// Deny login if user is deleted
-	if user.DeletedUser {
-		um.mu.RUnlock()
-		return nil, false
-	}
-	// Copy the password hash while holding the read lock
-	passwordHash := user.PasswordHash
-	um.mu.RUnlock()
-
-	// Compare hashed password outside any lock (bcrypt is CPU-intensive)
-	err := bcrypt.CompareHashAndPassword([]byte(passwordHash), []byte(password))
-	if err != nil {
-		return nil, false
-	}
-
-	// Authentication successful - update LastLogin and TimesCalled
-	um.mu.Lock()
-	user = um.users[lowerHandle] // Re-fetch under write lock
-	if user == nil {
-		um.mu.Unlock()
-		return nil, false
-	}
-	user.LastLogin = time.Now()
-	user.TimesCalled++
-	um.mu.Unlock()
-
-	// Save outside the write lock to avoid blocking other user operations
-	if err := um.SaveUsers(); err != nil {
-		slog.Error("failed to save user data after login", "handle", handle, "error", err)
-	}
-
-	// Return a copy
-	um.mu.RLock()
-	userCopy := *um.users[lowerHandle]
-	um.mu.RUnlock()
-	return &userCopy, true
-}
-
-// GetUser retrieves a user by handle (case-insensitive).
-// Returns a copy to prevent callers from mutating internal state without the lock.
-func (um *UserMgr) GetUser(handle string) (*User, bool) {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-
-	user, exists := um.users[strings.ToLower(handle)]
-	if !exists {
-		return nil, false
-	}
-	userCopy := *user
-	return &userCopy, true
-}
-
-// GetUserByID returns a user by their ID (for optimistic locking checks)
-func (um *UserMgr) GetUserByID(id int) (*User, bool) {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-
-	for _, user := range um.users {
-		if user.ID == id {
-			// Return a copy to prevent modification of the internal user data
-			userCopy := *user
-			return &userCopy, true
-		}
-	}
-	return nil, false
-}
-
-// NextUserID returns the ID that will be assigned to the next new user.
-func (um *UserMgr) NextUserID() int {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-	return um.nextUserID
-}
-
-// AddUser creates a new user, hashes the password, assigns an ID, and saves.
-func (um *UserMgr) AddUser(password, handle, realName, groupLocation string) (*User, error) {
-	handle = strings.TrimSpace(handle)
-	if handle == "" {
-		return nil, fmt.Errorf("handle cannot be blank")
-	}
-	lowerHandle := strings.ToLower(handle)
-
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Check if handle already exists
-	if _, exists := um.users[lowerHandle]; exists {
-		return nil, ErrHandleExists
-	}
-
-	// Hash the password
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		return nil, fmt.Errorf("failed to hash password: %w", err)
-	}
-
-	// Create new user
-	newUser := &User{
-		ID:            um.nextUserID,
-		PasswordHash:  string(hashedPassword),
-		Handle:        handle,
-		RealName:      realName,
-		GroupLocation: groupLocation,
-		AccessLevel:   um.newUserLevel,
-		TimeLimit:     60,
-		Validated:     false,
-		LastLogin:     time.Time{},
-	}
-
-	// Add to map and increment nextUserID
-	um.users[lowerHandle] = newUser
-	um.nextUserID++
-
-	// Save the updated user list *while still holding the lock*
-	if err := um.saveUsersLocked(); err != nil {
-		slog.Error("failed to save users after adding user", "handle", handle, "error", err)
-		delete(um.users, lowerHandle)
-		um.nextUserID--
-		return nil, err
-	}
-
-	slog.Info("added user", "handle", newUser.Handle, "id", newUser.ID)
-	return newUser, nil
-}
-
-// AddCallRecord adds a call record to the history and saves.
-func (um *UserMgr) AddCallRecord(record CallRecord) {
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Ensure slice exists
-	if um.callHistory == nil {
-		um.callHistory = make([]CallRecord, 0, callHistoryLimit)
-	}
-
-	// Assign the current call number and increment the counter
-	record.CallNumber = um.nextCallNumber
-	um.nextCallNumber++
-
-	// Append the new record
-	um.callHistory = append(um.callHistory, record)
-
-	// Limit the size of the history
-	if len(um.callHistory) > callHistoryLimit {
-		// Remove the oldest entry
-		um.callHistory = um.callHistory[1:]
-	}
-
-	// Save the updated history *while still holding the lock*
-	if err := um.saveCallHistoryLocked(); err != nil {
-		slog.Error("failed to save call history after adding record", "id", record.UserID, "error", err)
-		// Maybe try to rollback the append? Less critical than user add.
-	}
-}
-
-// GetLastCallers retrieves the recent call history (from memory).
-func (um *UserMgr) GetLastCallers() []CallRecord {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-
-	// Return a copy to prevent modification of the internal slice
-	historyCopy := make([]CallRecord, len(um.callHistory))
-	copy(historyCopy, um.callHistory)
-	return historyCopy
-}
-
-// GetAllUsers returns a slice containing copies of all user records.
-// Returns copies to prevent callers from mutating internal state.
-func (um *UserMgr) GetAllUsers() []*User {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-
-	usersSlice := make([]*User, 0, len(um.users))
-	for _, user := range um.users {
-		userCopy := *user
-		usersSlice = append(usersSlice, &userCopy)
-	}
-	return usersSlice
-}
-
-// GetUserCount returns the total number of registered users.
-func (um *UserMgr) GetUserCount() int {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-	return len(um.users)
-}
-
-// GetTotalCalls returns the total number of calls (logins) recorded.
-func (um *UserMgr) GetTotalCalls() uint64 {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-	if um.nextCallNumber <= 1 {
-		return 0
-	}
-	return um.nextCallNumber - 1
-}
-
-// SetNewUserLevel sets the access level assigned to new user signups.
-// This should be called after loading the server config.
-// Level is clamped to the valid range of 0-255.
-func (um *UserMgr) SetNewUserLevel(level int) {
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Validate and clamp to 0-255 range
-	if level < 0 {
-		slog.Warn("invalid newUserLevel; clamping to 0", "level", level)
-		level = 0
-	} else if level > 255 {
-		slog.Warn("invalid newUserLevel; clamping to 255", "level", level)
-		level = 255
-	}
-
-	um.newUserLevel = level
-}
-
-// MarkUserOnline marks a user as currently online/connected
-func (um *UserMgr) MarkUserOnline(userID int) {
-	um.mu.Lock()
-	defer um.mu.Unlock()
-	um.activeUserIDs[int32(userID)] = true
-	slog.Debug("user marked online", "id", userID)
-}
-
-// MarkUserOffline marks a user as offline/disconnected
-func (um *UserMgr) MarkUserOffline(userID int) {
-	um.mu.Lock()
-	defer um.mu.Unlock()
-	delete(um.activeUserIDs, int32(userID))
-	slog.Debug("user marked offline", "id", userID)
-}
-
-// IsUserOnline returns true if the user is currently connected
-func (um *UserMgr) IsUserOnline(userID int) bool {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-	return um.activeUserIDs[int32(userID)]
-}
-
-// PurgeResult holds information about a permanently purged user for reporting.
-type PurgeResult struct {
-	ID        int
-	Handle    string
-	DeletedAt time.Time
-}
-
-// PurgeDeletedUsers permanently removes soft-deleted users whose DeletedAt timestamp
-// is older than retentionDays days. Pass retentionDays=-1 to skip (no-op).
-// Returns a slice of PurgeResult describing the removed accounts.
-// The caller is responsible for logging admin activity if desired.
-func (um *UserMgr) PurgeDeletedUsers(retentionDays int) ([]PurgeResult, error) {
-	if retentionDays < 0 {
-		return nil, nil
-	}
-
-	cutoff := time.Now().AddDate(0, 0, -retentionDays)
-
-	um.mu.Lock()
-	defer um.mu.Unlock()
-
-	// Phase 1: identify eligible users without modifying the map yet.
-	type candidate struct {
-		key    string
-		user   *User
-		result PurgeResult
-	}
-	var candidates []candidate
-	for key, u := range um.users {
-		if !u.DeletedUser {
-			continue
-		}
-		if u.DeletedAt == nil {
-			// Deleted but no timestamp: treat as immediately eligible.
-			candidates = append(candidates, candidate{
-				key:  key,
-				user: u,
-				result: PurgeResult{
-					ID:     u.ID,
-					Handle: u.Handle,
-				},
-			})
-		} else if u.DeletedAt.Before(cutoff) {
-			candidates = append(candidates, candidate{
-				key:  key,
-				user: u,
-				result: PurgeResult{
-					ID:        u.ID,
-					Handle:    u.Handle,
-					DeletedAt: *u.DeletedAt,
-				},
-			})
-		}
-	}
-
-	if len(candidates) == 0 {
-		return nil, nil
-	}
-
-	// Phase 2: remove from in-memory store, then persist.
-	// Roll back in-memory changes if save fails so the store stays consistent.
-	for _, c := range candidates {
-		delete(um.users, c.key)
-	}
-	if err := um.saveUsersLocked(); err != nil {
-		for _, c := range candidates {
-			um.users[c.key] = c.user
-		}
-		return nil, fmt.Errorf("purge: failed to save users: %w", err)
-	}
-
-	purged := make([]PurgeResult, len(candidates))
-	for i, c := range candidates {
-		purged[i] = c.result
-	}
-	slog.Info("purged soft-deleted user accounts", "count", len(purged))
-	return purged, nil
-}
-
-// FindByAuthorizedKey returns the user whose registered PublicKeys include a
-// key matching the given marshaled wire bytes (ssh.PublicKey.Marshal()).
-// Matching is by exact key bytes; access-level authorization is enforced by
-// the caller, not here.
-func (um *UserMgr) FindByAuthorizedKey(marshaled []byte) (*User, bool) {
-	um.mu.RLock()
-	defer um.mu.RUnlock()
-	for _, u := range um.users { // um.users is map[string]*User
-		for _, line := range u.PublicKeys {
-			pub, _, _, _, err := ssh.ParseAuthorizedKey([]byte(line))
-			if err != nil {
-				continue
-			}
-			if string(pub.Marshal()) == string(marshaled) {
-				return u, true
-			}
-		}
-	}
-	return nil, false
 }
 
 // NewUserMgrForTest builds a UserMgr seeded with the given users, keyed by
